Add PoC for TestCase-less UnitTests
This allows – in combination with some small tweaks to phpunit itself –
to create Tests that are not dependent on the TestCase-Class.

The new tests need to implement the interface UnitTest which defines a
"setName" method as well s a constructor. Additionally a static
"getTestMethods" needs to be implemented to return all the methods that
shall be executed as tests. For convenience-Reasons there is an
implementation in a Trait "UnitTester" that can be added esaily.